fix(editor): fail closed on published records missing instructions too
Addresses CodeRabbit review on #21395:
- A stored/published record can exist with no `instructions` field at
  all (e.g. published before any were written). applyStoredOverrides
  now fails closed for that case too, not just unresolved/draft-only
  records.
- Moved the fail-closed regression tests to packages/editor/src/namespaces/agent.test.ts,
  colocated with the source they test.
- Added a before/after example to the changeset.
