+++ packages/editor/src/apply-stored-overrides.test.ts
@@ -464,79 +464,6 @@ describe('applyStoredOverrides', () => {
     expect(instructions).toBe('You are a code-defined agent.');
   });
 
-  // Regression tests for https://github.com/mastra-ai/mastra/issues/21373 —
-  // an agent with `editor: { instructions: true }` cannot provide code instructions
-  // (the type system forbids it), so if nothing resolves it must fail closed instead
-  // of silently generating with empty instructions.
-  describe('fails closed when editor exclusively owns instructions ("editor: { instructions: true }")', () => {
-    function makeEditorOwnedAgent() {
-      return new Agent({
-        id: 'editor-owned-agent',
-        name: 'Editor Owned Agent',
-        editor: { instructions: true, tools: false },
-        model: 'openai/gpt-4o',
-      });
-    }
-
-    it('throws when no stored agent record exists yet', async () => {
-      const storage = new InMemoryStore();
-      const editor = new MastraEditor();
-      const codeAgent = makeEditorOwnedAgent();
-      new Mastra({ storage, editor, agents: { 'editor-owned-agent': codeAgent } });
-
-      await expect(editor.agent.applyStoredOverrides(codeAgent, { status: 'published' })).rejects.toThrow(
-        /delegates instructions to the editor/,
-      );
-    });
-
-    it('throws when the stored agent is draft-only and status: "published" is requested', async () => {
-      const storage = new InMemoryStore();
-      const editor = new MastraEditor();
-      const codeAgent = makeEditorOwnedAgent();
-      new Mastra({ storage, editor, agents: { 'editor-owned-agent': codeAgent } });
-
-      const agentsStore = await storage.getStore('agents');
-      await agentsStore?.create({
-        agent: {
-          id: 'editor-owned-agent',
-          name: 'Editor Owned Agent',
-          instructions: 'DRAFT-ONLY-INSTRUCTIONS',
-          model: { provider: 'openai', name: 'gpt-4o' },
-        },
-        // no activeVersionId set -> draft-only, never published
-      } as Record<string, unknown>);
-
-      // Draft status still resolves normally.
-      const draftResolved = await editor.agent.applyStoredOverrides(codeAgent, { status: 'draft' });
-      expect(await draftResolved.getInstructions()).toBe('DRAFT-ONLY-INSTRUCTIONS');
-
-      // Published status has nothing to resolve — must fail closed.
-      await expect(editor.agent.applyStoredOverrides(codeAgent, { status: 'published' })).rejects.toThrow(
-        /no version has been published/,
-      );
-    });
-
-    it('throws when the storage adapter fails to load the stored config', async () => {
-      const storage = new InMemoryStore();
-      const editor = new MastraEditor();
-      const codeAgent = makeEditorOwnedAgent();
-      new Mastra({ storage, editor, agents: { 'editor-owned-agent': codeAgent } });
-
-      vi.spyOn(editor.agent as any, 'getStorageAdapter').mockRejectedValue(new Error('storage unavailable'));
-
-      await expect(editor.agent.applyStoredOverrides(codeAgent, { status: 'published' })).rejects.toThrow(
-        /delegates instructions to the editor/,
-      );
-    });
-
-    it('does not throw for a code-owned agent (no editor config) in the same unresolved scenarios', async () => {
-      // Sanity check: the fail-closed behavior is scoped to editor-owned instructions only.
-      const { editor, codeAgent } = await setup();
-      const result = await editor.agent.applyStoredOverrides(codeAgent, { status: 'published' });
-      expect(result).toBe(codeAgent);
-    });
-  });
-
   it('merges v1 toolProviders into the code agent tool list', async () => {
     const storage = new InMemoryStore();
     const codeTool = createTool({
